requirements.txt
requests>=2.28.0
pandas>=1.5.0
editdistance>=0.6.0
python-dotenv>=0.19.0 